Accept authorization cookie
If HTTP Authorization header is not present, check for a cookie named "authorization". The cookie's content is the same as what would be expected in the Authorization header.

This is to allow images to be retrieved via regular GET requets, as with <img> tags, without needing to make an Ajax request.