Implement gallery lightbox with infinite next-image advance
angular-bootstrap-lightbox is able to advance to the next or previous image with wrap-around (i.e. advancing pas the last image returns to the first image).

Since images in the gallery are progressively loaded based on whether the gallery is expanded, and whether the bottom of the gallery is visible, the default behaviour would have been to wrap around to the first image when advancing past the last already-loaded image, even if there are actually more images yet to be loaded from the server.

This custom implementation of the lightbox will, instead, behave as follows:

- When the gallery is expanded, advancing past the nth-last image (where n is the number of images that fit on one row) will trigger a pre-load of the next row, so the user can keep advancing as long as there are more images left on the server, without wrapping around to the first image

- When the gallery is collapsed, advancing past the last image will automatically expand the gallery and load the next row. Once the next row is loaded, the lightbox shows the first image from the next row. (Due to implementation details, there is an unavoidable flicker where the user will temporarily wrap-around to the first image before the correct ext image is shown.)