Backend implementation of invite requests
An invite request is a value in the state column of the subscription table.

When a user requests an invite to a stream, the coresponding row in the subscription table has its state column set to dbconst.sub.requested.

When a stream owner approes the invite request, the corresponding row in the subscription table has its state column set to dbconst.sub.invited.

Finally, when a user accepts an invite, the corresponding row in the subscription table has its state column set to dbconst.sub.active.

The following new APIs were implemented to facilitate the above functionality:
- GET api/invite/:id/request
  - id: stream ID
  - get list of invite requests for stream
- POST api/invite/:id/request
  - id: stream ID
  - userid (body): user ID
  - create invite request from user to stream
- DELETE api/invite/:id/request
  - id: stream ID
  - userid (query): user ID
  - delete invite request from user to stream
- GET api/subscription/:id/state
  - id: stream ID
  - userid (query): user ID
  - get state of user's subscription to stream (may be "active", "invited", "requested", or "none")