Gallery tracks oldest shown image by ID instead of time
The gallery previously sent the timestamp of the last image to be loaded in its call to /api/stream/contents, indicating that only images older than the given timestamp should be loaded. This could occasionally cause some images to be skipped if multiple imaeges were uploaded in the same second but not all of the could fit on the same row.

This problem can be avoided by sending the ID of the last image to be loaded instead, so the gallery always picks off from where it left off when more images are loaded.