POST api/image should only accept 1 stream ID
Following discussion, it was decided that allowing POST api/image to accept a comma-delimited list of stream IDs added undesirable complications to permissions testing. As a result, only one stream ID will be accepted. If an image needs to be associated with more than one stream, that can be accomplished with subsequent calls to POST api/image/:imageid/streams.

- Revised POST api/image to accept only 1 stream ID
- Added permission check to POST api/image to ensure user can only upload to a stream they own
- Added device.notify() to POST api/image/:imageid/streams to notify mobile devices when an image is added to a stream they're watching
- Upload page logic updated to accommodate the new way of uploading an image to multiple streams
- For legacy reasons, if an image upload fails due to the destination file already existing, the server response is { success: true }