Fix race condition in usermgr#/edit
When editing a user, there was a possibility that previously pre-filled form data could cause a cached user ID to be sent to the server, resulting in an erroneous success message when attempting to edit a user that doesn't actually exist.

The edit form now always queries the server for user info about the currently-entered username, and then uses the user ID from the response to that query.

Additionally, the pre-filled form data from a previous valid user lookup stayed in the form even after an invalid login was entered into the form. This has been corrected.