Fix collapsible alerts counter sometimes being wrong
This issue was being caused by an unnecessary kludge based on a misunderstanding of how ngAnimation works.

Previously, each alert had a "shown" property. ngIf would display a <uib-alert> element if alert.shown was truthy. When an lsAlerts.remove() was called, the alert's shown property would first be set to false to get the .ng-leave animation to trigger. A timer would be set so that when the animation is finished, the alert would be deleted from the model.

The timer to remove the alert from the model was set from inside lsAlerts.remove(). It therefore acted like a closure, inheriting its environment from lsAlerts.remove(). However, the environment was only evaluated when the timer expired; if lsAlerts.remove() was called again before th timer expired, the environment inherited by the timered function may change. This change in environment was occasionally causing the timered function to remove an alert from the wrong stack, resulting in the summary count on the collapse widget being incorrect (because any adjustments to the summary count are made _before_ the timer expires, not _as_ it expires).

But this was unnecessary; removing an an alert from the model removes it from the DOM, which is functionally identical to what ngIf does when the alert's shown property is set to false.